21.2 correct asumation, skip won games, tweaked this and that ...still doesn't work

+++ day_21_puzzle_02/src/die.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 
-pub fn dice_states() -> HashMap<u8, u32> {
+pub fn dice_states() -> HashMap<u16, u64> {
 	let mut states = HashMap::new();
 	
 	states.insert(3, 1);
